--- editor/source/actions.js
@@ -1,60 +1,0 @@
-function showLoading(flag) {
-    return {
-        type: 'SHOW_LOADING',
-        flag
-    }
-}
-
-function refreshList(data) {
-    return {
-        type: 'REFRESH_LIST',
-        data
-    }
-}
-
-export function fetchList() {
-    return dispatch => {
-        dispatch(showLoading(true));
-        fetch('http://localhost:8001/api/articles').then(function(response) {
-            return response.json();
-        }).then(function(data) {
-            dispatch(refreshList(data));
-            dispatch(showLoading(false));
-        }).catch(function(error) {
-            alert(JSON.stringify(error));
-        });
-    };
-}
-
-export function showList() {
-    return {
-        type: 'SHOW_LIST'
-    }
-}
-
-export function hideList() {
-    return {
-        type: 'HIDE_LIST'
-    }
-}
-
-function setContent(content) {
-    return {
-        type: 'SET_CONTENT',
-        content
-    }
-}
-
-export function openArticle(id) {
-    return dispatch => {
-        dispatch(showLoading(true));
-        fetch('http://localhost:8001/api/articles/' + id).then(function(response) {
-            return response.json();
-        }).then(function(data) {
-            dispatch(setContent(data));
-            dispatch(showLoading(false));
-        }).catch(function(error) {
-            alert(JSON.stringify(error));
-        });
-    };
-}
