+++ editor/source/actions/index.js
@@ -1,0 +1,3 @@
+import * as list from './list';
+
+export default list;
