+++ editor/source/components/editor.js
@@ -4,9 +4,17 @@ import ace from 'brace';
 import 'brace/mode/markdown';
 import 'brace/theme/tomorrow';
 
-export default class Editor extends React.Component {
-    componentWillReceiveProps (props) {
-        this.editor.setValue(props.content, -1);
+export default class extends React.Component {
+    resizeEditor () {
+        let width = window.innerWidth ||
+            document.documentElement.clientWidth ||
+            document.body.clientWidth;
+        if (width > 700) {
+            this.editor.renderer.setPadding((width - 700) / 2);
+        }
+    }
+    componentDidUpdate () {
+        this.editor.setValue(this.props.content, -1);
     }
     componentDidMount () {
         let editor = ace.edit('editor');
@@ -20,19 +28,14 @@ export default class Editor extends React.Component {
             fontSize: '14px',
             fontFamily: "Menlo, Consolas, 'source-code-pro', 'DejaVu Sans Mono', Monaco, 'Ubuntu Mono', 'Courier New', Courier, 'Microsoft Yahei', 'Hiragino Sans GB', 'WenQuanYi Micro Hei', monospace",
             hScrollBarAlwaysVisible: false,
+            selectionStyle: "line",
             displayIndentGuides: false
         });
         editor.renderer.setScrollMargin(200, 200);
-        window.addEventListener('resize', function() {
-            let width = window.innerWidth ||
-                document.documentElement.clientWidth ||
-                document.body.clientWidth;
-            if (width > 700) {
-                editor.renderer.setPadding((width - 700) / 2);
-            }
-        });
         editor.container.style.lineHeight = 2;
         this.editor = editor;
+        this.resizeEditor();
+        window.addEventListener('resize', this.resizeEditor.bind(this));
     }
     render() {
         return (
