+++ editor/source/components/search.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import classNames from 'classnames';
+
+export default class extends React.Component {
+    render() {
+        const list = this.props.list;
+        const actions = this.props.actions;
+        return (
+            <label id="search-wrap" htmlFor="search">
+                <i className="fa fa-search"></i>
+                <input id="search" type="text" placeholder="搜索..." />
+            </label>
+        );
+    }
+}
