--- editor/source/stores/article.js
@@ -1,2 +1,0 @@
-import { createStore } from 'redux';
-export default {};
