+++ editor/source/styles/index.less
@@ -3,7 +3,7 @@
 .hover {
     cursor: pointer;
     &:hover {
-        opacity: 0.7;
+        opacity: 0.9;
     }
 }
 
@@ -29,6 +29,7 @@
 }
 #left {
     padding: 30px;
+    min-width: 42px;
     position: absolute;
     left: 0;
     top: 0;
@@ -36,9 +37,12 @@
     z-index: 2;
     overflow: auto;
     border-right: 1px dashed #ddd;
-    // background-color: white;
+    background-color: white;
+    background-color: #fefefe;
+    .vendor(transition, width 0.25s);
     .menu {
-        float: left;
+        // float: left;
+        position: fixed;
         width: 40px;
         .button {
             margin: 0 auto;
@@ -53,6 +57,67 @@
             }
             &.setting {
                 margin-top: 15px;
+            }
+        }
+    }
+}
+#files {
+    width: 250px;
+    margin-left: 40px;
+    padding: 13px 30px;
+    padding-top: 0;
+    padding-right: 0;
+    #search-wrap {
+        position: fixed;
+        top: 0;
+        rigth: 0;
+        z-index: 2;
+        padding-top: 44px;
+        padding-bottom: 30px;
+        color: #aaa;
+        background-image: linear-gradient(to top, transparent, white 50%);
+        .fa {
+            position: absolute;
+            margin-top: 0;
+            font-size: 17px;
+        }
+        #search {
+            padding-left: 25px;
+            border: 0;
+            display: block;
+            width: 100%;
+            background-color: transparent;
+        }
+    }
+    #list {
+        // display: none;
+        margin-top: 70px;
+        .item {
+            padding-bottom: 40px;
+            &.selected {
+                .title {
+                    color: #000;
+                    border-left: 4px solid #702500;
+                }
+            }
+            .head {
+                .date {
+                    color: #aaa;
+                }
+            }
+            .title {
+                color: #777;
+                font-weight: bold;
+                font-size: 16px;
+                margin-top: 10px;
+                margin-left: -13px;
+                padding-left: 9px;
+                border-left: 4px solid transparent;
+                .vendor(box-sizing, border-box);
+            }
+            .preview {
+                color: #bbb;
+                margin-top: 10px;
             }
         }
     }
@@ -78,63 +143,16 @@
         }
     }
 }
-
-#files {
-    width: 200px;
-    float: left;
-    padding: 13px 30px;
-    #search-wrap {
-        display: block;
-        width: 70%;
-        color: #aaa;
-        .fa {
-            position: absolute;
-            margin-top: 0;
-            font-size: 17px;
-        }
-        #search {
-            padding-left: 25px;
-            border: 0;
-            display: block;
-            width: 100%;
-            background-color: transparent;
-        }
-    }
-    #list {
-        // display: none;
-        margin-top: 30px;
-        .item {
-            // border-bottom: 1px solid #eee;
-            padding-bottom: 30px;
-            .head {
-                margin-bottom: 5px;
-                .date {
-                    color: #aaa;
-                }
-            }
-            .title {
-                font-weight: bold;
-                margin-bottom: 5px;
-                font-size: 16px;
-            }
-            .preview {
-                color: #bbb;
-                // font-style: italic;
-            }
-        }
-    }
-}
 #header {
     position: absolute;
     left: 0;
     right: 0;
     top: 0;
     margin: 0 20px;
-    padding-top: 50px;
+    padding-top: 80px;
     padding-bottom: 50px;
     z-index: 1;
     text-align: center;
-    // background-color: white;
     pointer-events: none;
     background-image: linear-gradient(to top, transparent, white 50%);
     .title {
